# swagger_docs.yaml
openapi: 3.1.0
info:
  title: 'Hello API Platform'
  description: ''
  version: 1.0.0
servers:
  -
    url: /
    description: ''
paths:
  /api/items:
    get:
      operationId: api_items_get_collection
      tags:
        - Item
      responses:
        200:
          description: 'Item collection'
          content:
            application/ld+json:
              schema:
                type: object
                properties:
                  'hydra:member': { type: array, items: { $ref: '#/components/schemas/Item.jsonld' } }
                  'hydra:totalItems': { type: integer, minimum: 0 }
                  'hydra:view': { type: object, properties: { '@id': { type: string, format: iri-reference }, '@type': { type: string }, 'hydra:first': { type: string, format: iri-reference }, 'hydra:last': { type: string, format: iri-reference }, 'hydra:previous': { type: string, format: iri-reference }, 'hydra:next': { type: string, format: iri-reference } }, example: { '@id': string, type: string, 'hydra:first': string, 'hydra:last': string, 'hydra:previous': string, 'hydra:next': string } }
                  'hydra:search': { type: object, properties: { '@type': { type: string }, 'hydra:template': { type: string }, 'hydra:variableRepresentation': { type: string }, 'hydra:mapping': { type: array, items: { type: object, properties: { '@type': { type: string }, variable: { type: string }, property: { type: [string, 'null'] }, required: { type: boolean } } } } } }
                required:
                  - 'hydra:member'
      summary: 'Retrieves the collection of Item resources.'
      description: 'Retrieves the collection of Item resources.'
      parameters:
        -
          name: page
          in: query
          description: 'The collection page number'
          required: false
          deprecated: false
          allowEmptyValue: true
          schema:
            type: integer
            default: 1
          style: form
          explode: false
          allowReserved: false
      deprecated: false
    post:
      operationId: api_items_post
      tags:
        - Item
      responses:
        201:
          description: 'Item resource created'
          content:
            application/ld+json:
              schema:
                $ref: '#/components/schemas/Item.jsonld'
          links: {  }
        400:
          description: 'Invalid input'
        422:
          description: 'Unprocessable entity'
      summary: 'Creates a Item resource.'
      description: 'Creates a Item resource.'
      parameters: []
      requestBody:
        description: 'The new Item resource'
        content:
          application/ld+json:
            schema:
              $ref: '#/components/schemas/Item.jsonld'
        required: true
      deprecated: false
    parameters: []
...
components:
  schemas:
...
    Item.jsonld:
      type: object
      description: ''
      deprecated: false
      required:
        - creator
        - name
        - isValidated
        - isArchived
      properties:
        '@context':
          readOnly: true
          oneOf:
            -
              type: string
            -
              type: object
              properties:
                '@vocab':
                  type: string
                hydra:
                  type: string
                  enum: ['http://www.w3.org/ns/hydra/core#']
              required:
                - '@vocab'
                - hydra
              additionalProperties: true
        '@id':
          readOnly: true
          type: string
        '@type':
          readOnly: true
          type: string
        creator:
          'owl:maxCardinality': 1
          type:
            - string
            - 'null'
          format: iri-reference
          example: 'https://example.com/'
        validator:
          'owl:maxCardinality': 1
          type:
            - string
            - 'null'
          format: iri-reference
          example: 'https://example.com/'
        tags:
          type: array
          items:
            type: string
            format: iri-reference
            example: 'https://example.com/'
        id:
          readOnly: true
          type: integer
        name:
          type: string
        publication_date:
          type:
            - string
            - 'null'
          format: date-time
        description:
          minLength: 20
          type: string
        validation_date:
          type:
            - string
            - 'null'
          format: date-time
        isValidated:
          type: boolean
        isArchived:
          type: boolean
        publicationDate:
          type:
            - string
            - 'null'
          format: date-time
        validationDate:
          type:
            - string
            - 'null'
          format: date-time
...
  securitySchemes: {  }
security: []
...